Create CityStateMap.java
This mapping class will allow the user to add new city-state pair in to it. User can get the state of the city by calling get method on this class. Duplicate city entries are not allowed,  in such case previous mapping should be replaced with new one. Assume that there is a text file having city state values (in string).